# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: School Bus Tracking API
  version: 0.7.0
  description: |
    API for the School Bus Tracking admin web and the Android Driver and Parent apps.

    ## Conventions
    - Base path for business endpoints: `/api/v1`. JSON fields are `snake_case`.
    - Timestamps are RFC 3339 UTC (`2026-10-07T02:15:00Z`); calendar dates are `YYYY-MM-DD`.
    - Mobile numbers are accepted as `9876543210`, `+91 98765 43210`, etc. and always returned as `+919876543210`.
    - Success responses wrap the payload: `{"data": ...}`. Lists add
      `"meta": {"page", "page_size", "total"}` and accept `?page=1&page_size=20` (max 100).
    - Errors: `{"error": {"code", "message", "fields"}}`. Branch on `code`, show `message`.
    - Every response carries an `X-Request-ID` header; include it when reporting a bug.

    ## Authentication
    - **Driver and Parent apps:** `POST /auth/otp/send` then `POST /auth/otp/verify` with `app` = `driver` or `parent`.
      On the dev environment every code is `123456` (no SMS is sent).
    - **Admin web:** `POST /auth/login` (email + password). If the response has `mfa_required: true`,
      call `POST /auth/login/2fa` with the `mfa_token` and the authenticator code.
    - Send `Authorization: Bearer <access_token>` on every other call.
    - Access tokens last `expires_in` seconds (15 min). On `401`, call `POST /auth/refresh` once with the
...
      latest one. If refresh fails, send the user to login.
...

...

    ### Driver app: sending GPS
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/schools/{schoolID}/routes:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [Routes & stops]
      summary: List routes
      description: Without stops; each route has `stop_count`.
      operationId: listRoutes
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - { name: q, in: query, description: Search name or code, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive] } }
      responses:
        "200":
          description: Routes, by code.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Route" } }
                  meta: { $ref: "#/components/schemas/PageMeta" }
    post:
      tags: [Routes & stops]
      summary: Create a route
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RouteInput" }
      responses:
        "201":
          description: Created (with an empty `stops` list).
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Route" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/routes/{routeID}:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/RouteID" }
    get:
      tags: [Routes & stops]
      summary: Get a route with its stops in order
      operationId: getRoute
      responses:
        "200":
          description: The route; `stops` sorted by `sequence`.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Route" } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Routes & stops]
      summary: Update a route
      operationId: updateRoute
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RouteInput" }
      responses:
        "200":
          description: Updated (without stops).
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Route" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/routes/{routeID}/status:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/RouteID" }
    patch:
      tags: [Routes & stops]
      summary: Activate or deactivate a route
      operationId: setRouteStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { type: string, enum: [active, inactive] } }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Route" } } }
  /api/v1/schools/{schoolID}/routes/{routeID}/stops:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/RouteID" }
    post:
      tags: [Routes & stops]
      summary: Add a stop
      description: Appends to the end, or inserts at `position` and shifts later stops down.
      operationId: createStop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/StopInput"
                - type: object
                  properties:
                    position: { type: integer, minimum: 1, description: "1-based; omit to append." }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Stop" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/schools/{schoolID}/routes/{routeID}/stops/order:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/RouteID" }
    put:
      tags: [Routes & stops]
      summary: Reorder stops
      description: Send every stop ID of the route, once each, in the new travel order.
      operationId: reorderStops
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stop_ids]
              properties:
                stop_ids: { type: array, items: { type: string, format: uuid } }
      responses:
        "200":
          description: Stops in the new order.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Stop" } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/schools/{schoolID}/routes/{routeID}/stops/{stopID}:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/RouteID" }
      - { name: stopID, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Routes & stops]
      summary: Update a stop
      description: Position is not changed here; use the reorder endpoint.
      operationId: updateStop
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StopInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Stop" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Routes & stops]
      summary: Delete a stop
      description: Later stops move up so the sequence stays 1..N.
      operationId: deleteStop
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SchoolID: { name: schoolID, in: path, required: true, schema: { type: string, format: uuid } }
    RouteID: { name: routeID, in: path, required: true, schema: { type: string, format: uuid } }
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  schemas:
...
    RouteInput:
      type: object
      required: [name, code]
      properties:
        name: { type: string }
        code: { type: string, example: RS-01, description: Unique per school; stored upper-case. }
        start_point: { type: string, example: School }
        description: { type: string }
        supports_pickup: { type: boolean, default: true, description: Used for Morning Pickup trips. }
        supports_drop: { type: boolean, default: true, description: Used for Evening Drop trips. }
    Route:
      allOf:
        - $ref: "#/components/schemas/RouteInput"
        - type: object
          required: [id, school_id, status, stop_count, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            school_id: { type: string, format: uuid }
            status: { type: string, enum: [active, inactive] }
            stop_count: { type: integer }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            stops:
              type: array
              items: { $ref: "#/components/schemas/Stop" }
              description: Only on GET of a single route, ordered by sequence.
    StopInput:
      type: object
      required: [name, latitude, longitude]
      properties:
        name: { type: string, example: Gandhipuram }
        landmark: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90, example: 11.0168 }
        longitude: { type: number, minimum: -180, maximum: 180, example: 76.9663 }
        pickup_time:
          type: [string, "null"]
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "07:10"
          description: Scheduled Morning Pickup time (HH:MM).
        drop_time:
          type: [string, "null"]
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "16:40"
          description: Scheduled Evening Drop time (HH:MM).
        geofence_radius_m: { type: integer, minimum: 25, maximum: 1000, default: 100, description: Radius for Reached detection. }
    Stop:
      allOf:
        - $ref: "#/components/schemas/StopInput"
        - type: object
          required: [id, route_id, sequence, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            route_id: { type: string, format: uuid }
            sequence: { type: integer, minimum: 1, description: "Travel order, 1..N without gaps." }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
  responses:
...
    ValidationFailed:
      description: Invalid input.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: validation_failed, message: One or more fields are invalid., fields: { mobile: must be a valid 10-digit mobile number } } }
...
    NotFound:
      description: Missing, or belongs to another school.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Duplicate value.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: conflict, message: A record with this email already exists., fields: { email: already exists } } }
...